Объявления отрисовываются на карте в соответствии с данными с сервера.
